+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ### Ver 0.1.1(Not released)
 
+  * #1 [Improve tab title color](https://github.com/georgeOsdDev/react-draggable-tab/issues/1)
+
 ### Ver 0.1.0 Initial release